requirements.txt
flake8
python-dotenv
Flask
Flask-sqlalchemy
Flask-WTF
Flask-Bootstrap4
Flask-Session
Flask-Testing
spotipy == 2.16.1